proj1/consumer/requirements.txt
dnspython==2.4.2
kafka-python==2.0.2
pymongo==4.6.1
